requirements.txt
python-telegram-bot==10.0.1
feedparser==5.2.1
python-dateutil=2.7.2
